fix (#11)
* feat: Feed Director

* fix: fix

* feat: creation of functionality

* fix: int to long

* fix: int to long

* fix: update readme

* fix: update readme

* fix: update readme

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix

* fix: изменение типа данных в отзывах с long на Long

* fix: изменение типа данных в отзывах с long на Long

* fix: FilmStorageTest

* fix: GetGenreByIdTest

* fix: GetGenreByIdTest

* fix: UserStorageTest

* fix: UserStorageTest

* fix: UserStorageTest

* fix: UserStorageTest

---------

+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/impl/UserDbStorage.java
@@ -71,7 +71,7 @@ public class UserDbStorage implements UserStorage {
             statement.setDate(4, Date.valueOf(user.getBirthday()));
             return statement;
         }, keyHolder);
-        user.setId(keyHolder.getKey().intValue());
+        user.setId(keyHolder.getKey().longValue());
         return user;
     }
